requirements.txt
Flask>=2.0.0
gunicorn>=20.1.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.0
scikit-learn>=1.0.0
joblib>=1.1.0


